requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.0
Flask-SocketIO==5.5.1
gunicorn==23.0.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pyspellchecker==0.8.3
python-engineio==4.12.1
python-socketio==5.13.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
